Changed method to GET. Moved the EventMapData class to another file.
