+++ src/main/java/model/EventMapData.java
@@ -1,0 +1,15 @@
+package model;
+
+public class EventMapData {
+    private String id;
+    private double longitude;
+    private double latitude;
+    private String name;
+
+    public EventMapData(String id, double longitude, double latitude, String name) {
+        this.id = id;
+        this.longitude = longitude;
+        this.latitude = latitude;
+        this.name = name;
+    }
+}
