--- src/main/java/model/EventsMapRequest.java
@@ -1,36 +1,0 @@
-package model;
-
-import org.codehaus.jackson.annotate.JsonProperty;
-
-public class EventsMapRequest {
-    @JsonProperty("longitude")
-    private double longitude;
-    @JsonProperty("latitude")
-    private double latitude;
-    @JsonProperty("radius")
-    private double radius;
-
-    public double getLongitude() {
-        return longitude;
-    }
-
-    public void setLongitude(double longitude) {
-        this.longitude = longitude;
-    }
-
-    public double getLatitude() {
-        return latitude;
-    }
-
-    public void setLatitude(double latitude) {
-        this.latitude = latitude;
-    }
-
-    public double getRadius() {
-        return radius;
-    }
-
-    public void setRadius(double radius) {
-        this.radius = radius;
-    }
-}
